requirements.txt
fastapi
uvicorn

opencv-python
numpy
omegaconf
pydantic

# После установки пакета в файле compute.py заменить from skipy.ndimage на from skimage.io
cpbd

# Пакеты для тестов, линтера, ..
httpx
pytest
wemake-python-styleguide

realesrgan
flask
pdf2image
pypdf2
requests
pillow